Seamonkey version bump to 2.30
comm-release for mozilla-33 does not build properly when MOZ_OBJDIR is set, without
running the configure phase twice -- it ignores .mozconfig the first time around;
this ebuild does that, as ugly as it is.

Also applied the jemalloc configure fix needed to fix bug 525522, to seamonkey.

Also in this commit is an explicit flag for system bzip2, and a drop of a redundant
--with-system-zlib from mozconfig-v5.33.eclass.  An initial build test was failing
on the bundled bzip2, so this just avoids that.