Add firefox-34.0_beta2, provide alternative to binary-fetched gmp
Firefox-33 and above introduced the Gecko Media Plugin infrastructure and
its first plugin (Cisco's BSD-licensed openh264 codec) for webrtc h264 support.
By default, and due to software patent issues related to h264, Cisco builds
the plugin themselves (so that they are on the hook rather than end-users) and
firefox updates this non-optional plugin via downloading the binary blob into
each user's profile.  To help resolve bug 525810, firefox-33 and firefox-34_beta2
introduce USE="system-gmps" which will disable the autoupdates of gmp plugins in
the firefox default prefs, and RDEPEND on the newly added media-plugins/gmp-openh264.

The gmp-openh264 package builds the plugin from source and installs it to the
regular nsplugins path, in a directory tree structure required for GMP.  It also
adds an env.d file to set the MOZ_GMP_PATH environment variable appropriately, so that
firefox will load the plugin from here if it cannot be found in the user's profile.

USE="system-gmps" plus media-plugins/gmp-openh264 will:

1 - make the system-installed plugin always be used on newly-created user profiles
or updated old profiles from firefox-32 and earlier.

2 - allow users to obtain the mozilla/cisco binary simply by enabling auto-udates
in about:config or manually checking for and installing updates via the Addon Manager

3 - *not* override a user's in-profile copy of the mozilla/cisco binary if it exists;
this includes if the user has manually fetched it via overriding auto-updates or doing
a manual update in Addon Manager, and also (unfortunately) if the user has already
installed and run firefox-33.0.

This will remain an experiment in mozilla-overlay for a while, as maintenance
of an openh264 may not prove to be viable -- at this point, it looks like we will need to
roll our own snapshots of two github upstream projects in order to provide release
tarballs, and it is difficult to say how many releases there will need to be going forward.