Bumped firefox-34 to beta6, bumped gmp-openh264 and adjusted system-gmps override
The version string as well as the path reported in about:plugins and about:addons
is faked entirely by the media.gmp-gmpopenh264.version pref; as well this pref is
used to determine the 'installed' status of the plugin.  This bump adds a fake and static
.version pref to show the plugin is 'system-installed', which allows the system plugin
to be updated without a need for the user to refresh their login session and also
show the plugin as being installed rather than 'will be installed shortly'.