Thunderbird version bump, integrating some changes and improving locales support
Includes an example 38.0.1 ebuild that builds localizations from l10n sources, including
for lightning and gdata-provider

The 38.1.0 bump uses pre-packaged localizations but leverages the new helpers from mozlinguas.eclass.
It also repurposes USE="lightning" to install lightning and gdata-provider in app-global fashion, the
same as it was in thunderbird-31.x.  FYI, this use flag might not remain by the time the package hits
the portage tree.