minor fixes to mozlinguas.eclass
First, dropped any form of inherit on mozcoreconf-v3.  Instead,
mozlinguas_mozconfig will error if mozconfig_annotate isn't declared in the
very rare cases that it is actually needed.  It has to be anyways, otherwise
the .mozconfig file itself isn't being eclass managed in the first place.

Second, ensured the target dir for the additional locales for extensions
will exist, thus allowing the function to be used in thunderbird-bin
for installing alt locales into its bundled lightning (and likely others).