unset MOZ_DEV_EDITION define but add back some of the pref's when USE=bindist
Since firefox-35, USE=bindist has enabled codepaths that completely change how profiles
are handled, as per bug 536682.  This patch attempts to address that by disabling this
codepath, but keeping some of the pref() entries that the codepath enables.  The pref()
entries may not necessarily be a good idea to keep either, but we may not want a dev-edition
branded firefox to act exactly like a firefox-branded one either (in terms of crashreporting,
feedback, and so forth).