cleaned old, synced 41 with the main tree, and introduced 42
Firefox-42 introduces a new base directory -- gone is the
'mozilla-{beta,release,esr}' pattern, and now the base source dir essentially
matches the distfile's name.  it is unclear how this will pan out for ESR's
so we will need to watch out.

While inspecting dependency versions I discovered that what we are depending on
for the necko-wifi module needs some work.  The implementation requires dbus
for linux, because the only querying the module does is against
'org.freedesktop.NetworkManager', meaning the operational dep is actually NM
rather than wireless-tools.  It also seems that support on other platforms
(bsd etc) is done more or less natively, so i wrapped the deps with 'kernel_linux'
to address that.