bump firefox-46 to beta6, put back optional jit
Testing on hardened showed that despite upstream's declaration that all RWX bits of code
are now W^X, there are still a number of mprotect related errors.  Putting back the jit flag
so that it can be disabled for hardened, and reintroducing the various patches etc to support
non-jit builds.