www-client/seamonkey-2.42.2.0_p0 - build seamonkey out of thunderbird-45 distfiles
Seamonkey upstream is still having large difficulties in rolling releases, so
its time to take matters into our own hands.  Packages suffixed with a _p0 version
are packages that will be built out of thunderbird's distfiles, with the added
extensions that are not normally included in the thunderbird codebase.

Vesion mapping is as follows -- seamonkey-2.42 maps to mozilla45 core, so we substitute
mozilla's major version number for seamonkey's major.minor in the version number for
thunderbird.  Therefore, seamonkey-2.42.2.0 uses the sources also used to build
thunderbird-45.2.0

I should have no problem keeping up with the thunderbird-45 release schedule for
the remainder of mozilla-45-ESR, as long as this package works as expected for end users.