firefox-48 - use skia by default
Skia is the default upstream 2D canvas renderer, and it is highly recommended
though not yet fully required to use it.  This commit enables it as a USE flag
(default-enabled), and in order to circumvent a bug, forces a pref to the
value 'cairo' (instead of the erroneous default value 'skia') when the flag
is disabled.