www-client/firefox: bump to 52.0.2, add apulse support
This package in addition to bumping, does some pkg_preinst() magic
to support using apulse in place of a full pulseaudio setup when
USE="pulseaudio" is specified.  This will likely not go into the
gentoo repo until such time as alsa support is broken upstream.

This bump also includes a patch that should help compilation on
musl.