Bump seamonkey to 2.49.2 out of thunderbird-52.2 sources
This package builds localization directly from sources.  Locales are
patched to support the latest chatzilla as a bug keeps incomplete locales
from building.

The plan going forward is to keep up with security patching by bumping
seamonkey in this manner for every thunderbird release, at least for the
remainder of the 52.x series.