www-client/firefox: add eme-free use flag to buildtime-disable all DRM plugin loading
For those wanting to ensure firefox will never load any closed DRM content plugins
or allow any user on the system to do so in their profile, USE="eme-free" now
triggers --disable-eme in the build configuration, which entirely removes the
capability of DRM plugin loading.  This flag nullifies the effect that
USE=-gmp-autoupdate performs but a REQUIRED_USE was not implemented as
forcing package.use modifications seems overkill for what becomes a do-nothing
flag.