seamonkey: bump to 2.49 built from comm-esr52.4.0
There is an upstream release candidate 2.49.1 (build2) that is checked out
from the same upstream codebase; this package uses the officially released
thunderbird tarball instead but leverages the localization files from the
release candidate.

This ebuild may go away in favour of a direct ebuild to upstream's 2.49.1
if upstream releases it in a timely fashion (at present the release is
waiting on build issues of non-linux platforms).