eclass/mozcoreconf-v6.eclass: adjust for PYTHON3 support in ebuilds
As python2 only has one slot, but python3 support is required for v61
and newer AND may be different from one mozilla release to another,
this commit honours the setting of PYTHON_COMPAT and PYTHON_REQ_USE
in the header of ebuilds so that python3 variants can be chosen
via python-any-r1.  If not specified, the previous python2-only
values are set for compatibility.

If a python3 variant is specified in PYTHON_COMPAT, then
after python-any-r1 has chosen the correct python variant to use
this value is set in the PYTHON3 environment variable, and then
PYTHON itself is forced back to python2 (since mozilla requires
both to build for the forseeable future)