thunderbird: attempt to fix setupterm issues by patch
The python lib 'blessings' is what requires setupterm, and it
seems to be semi-optional based on the code that loads it; this
commit drops the need for blessings by forcing a return of
NullTerminal instead of only doing so when blessings.Terminal throws
an exception.

Note - this is a debug commit, the final fix will involve properly
running down why the exception isn't caught, or how to force
the disable_colors setting to ensure NullTerminal is always used.