thunderbird: bump to 60.3.0
Stopped running the lightning version check, as the relevance of
the response is not actually useful -- the reported lightning
version is calculated directly from the provided ${PV} and has
nothing to do with the changes to lightning itself.  Leaving the code
there, commented out, just in case it's still useful in the future.